test: Add socket.resume() calls to trigger old mode
Many of our socket tests don't actually add a 'data' listener, but they rely
on old-mode semantics (which we definitely want to keep testing).

Trigger the old mode by calling resume().  Effectively, it's as if streams
start out in a 'paused' state, but then 'unpause' if you add a data listener.